Fix hoisting errors and refactor ProgressionTimingIndicator component
